Cleanup Twizzler crate.
