+++ src/lib/twizzler/src/lib.rs
@@ -15,7 +15,6 @@ pub mod collections;
 pub mod marker;
 pub mod object;
 pub mod ptr;
-pub mod tx;
 
 pub(crate) mod util;
