+++ src/lib/twizzler/src/ptr/resolved/resolved_tx_slice.rs
@@ -1,114 +1,12 @@
 use std::{
     borrow::{Borrow, BorrowMut},
-    mem::MaybeUninit,
     ops::{Deref, DerefMut, Index, IndexMut, RangeBounds},
 };
 
 use twizzler_rt_abi::object::ObjectHandle;
 
-use super::TxObject;
-use crate::{
-    object::RawObject,
-    ptr::{GlobalPtr, Ref, RefMut},
-    util::range_bounds_to_start_and_end,
-};
-
-pub struct TxRef<T> {
-    ptr: *mut T,
-    tx: Option<TxObject<()>>,
-}
-
-impl<T> TxRef<T> {
-    pub fn as_mut(&mut self) -> RefMut<'_, T> {
-        let handle = self.tx.as_ref().unwrap().handle().handle();
-        unsafe { RefMut::from_raw_parts(self.ptr, handle) }
-    }
-
-    pub unsafe fn from_raw_parts<B>(tx: TxObject<B>, ptr: *mut T) -> Self {
-        Self {
-            ptr,
-            tx: Some(tx.into_unit()),
-        }
-    }
-
-    #[inline]
-    pub fn offset(&self) -> u64 {
-        self.handle().ptr_local(self.ptr.cast()).unwrap() as u64
-    }
-
-    pub fn tx(&self) -> &TxObject<()> {
-        self.tx.as_ref().unwrap()
-    }
-
-    pub fn tx_mut(&mut self) -> &mut TxObject<()> {
-        self.tx.as_mut().unwrap()
-    }
-
-    pub fn into_tx(mut self) -> TxObject<()> {
-        self.tx.take().unwrap()
-    }
-
-    pub fn raw(&self) -> *mut T {
-        self.ptr
-    }
-
-    pub fn handle(&self) -> &ObjectHandle {
-        self.tx().handle()
-    }
-
-    pub fn global(&self) -> GlobalPtr<T> {
-        GlobalPtr::new(self.handle().id(), self.offset())
-    }
-}
-
-impl<T> TxRef<MaybeUninit<T>> {
-    pub fn write(mut self, val: T) -> crate::Result<TxRef<T>> {
-        unsafe {
-            let ptr = self.ptr.as_mut().unwrap_unchecked();
-            let tx = self.tx.take().unwrap();
-            Ok(TxRef::<T>::from_raw_parts(tx, ptr.write(val)))
-        }
-    }
-}
-
-impl<T> Deref for TxRef<T> {
-    type Target = T;
-
-    fn deref(&self) -> &Self::Target {
-        unsafe { self.ptr.as_ref().unwrap_unchecked() }
-    }
-}
-
-impl<T> DerefMut for TxRef<T> {
-    fn deref_mut(&mut self) -> &mut Self::Target {
-        unsafe { self.ptr.as_mut().unwrap_unchecked() }
-    }
-}
-
-impl<T> Drop for TxRef<T> {
-    #[track_caller]
-    fn drop(&mut self) {
-        let _ = self.tx.take().map(|mut tx| tx.commit());
-    }
-}
-
-impl<T> Into<ObjectHandle> for TxRef<T> {
-    fn into(self) -> ObjectHandle {
-        self.tx().handle().clone()
-    }
-}
-
-impl<T> Into<ObjectHandle> for &TxRef<T> {
-    fn into(self) -> ObjectHandle {
-        self.tx().handle().clone()
-    }
-}
-
-impl<T> AsRef<ObjectHandle> for TxRef<T> {
-    fn as_ref(&self) -> &ObjectHandle {
-        self.tx().handle()
-    }
-}
+use super::{Ref, RefMut, TxRef};
+use crate::{ptr::GlobalPtr, util::range_bounds_to_start_and_end};
 
 pub struct TxRefSlice<T> {
     ptr: TxRef<T>,
@@ -118,6 +16,11 @@ pub struct TxRefSlice<T> {
 impl<T> TxRefSlice<T> {
     pub unsafe fn from_ref(ptr: TxRef<T>, len: usize) -> Self {
         Self { ptr, len }
+    }
+
+    #[inline]
+    pub fn offset(&self) -> u64 {
+        self.ptr.offset()
     }
 
     pub fn as_slice(&self) -> &[T] {
@@ -168,6 +71,12 @@ impl<T> TxRefSlice<T> {
         } else {
             unsafe { Self::from_ref(self.ptr, 0) }
         }
+    }
+}
+
+impl<T> From<TxRefSlice<T>> for GlobalPtr<T> {
+    fn from(value: TxRefSlice<T>) -> Self {
+        GlobalPtr::new(value.handle().id(), value.offset())
     }
 }
