--- src/lib/twizzler/src/tx.rs
@@ -1,5 +1,0 @@
-mod object;
-mod reference;
-
-pub use object::*;
-pub use reference::*;
